Mangle names of synthetic script classes
A script has to be compiled into a class to be executed, but it is
feasible to create a class within a script that has the same name of the
script file. This creates a conflict as the class and the script would
end up being compiled with the same name. Mangling the script name
ensures this cannot happen.

The mangling is kept to a minimum, by adding a $ at the end of the
generated script class name.

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>